+++ apps/api/src/app/converting/converting.module.ts
@@ -1,0 +1,11 @@
+import { Module } from "@nestjs/common";
+import { SetsModule } from "../sets/sets.module";
+import { UsersModule } from "../users/users.module";
+import { CardsModule } from "../cards/cards.module";
+import { ConvertingController } from "./converting.controller";
+
+@Module({
+  imports: [SetsModule, UsersModule, CardsModule],
+  controllers: [ConvertingController]
+})
+export class ConvertingModule {}
