+++ apps/api/src/app/shared/response/error.response.ts
@@ -2,7 +2,7 @@ import { ApiProperty } from "@nestjs/swagger";
 
 export class ErrorResponse {
   @ApiProperty({
-    description: "Denotes whether the request was successful or not",
+    description: "Denotes whether the server or client is at fault in the error. \"error\" indicates that the server is at fault, while \"fail\" indicates that the client is at fault.",
     example: "fail"
   })
     status: string;
