+++ apps/front/src/app/shared/http/sets.service.ts
@@ -71,91 +71,6 @@ export class SetsService {
   }
 
   /**
-   * Converts a set to a CSV file
-   *
-   * @param setId ID of the set to convert
-   *
-   * @returns Blob of the .csv
-   */
-  async convertSetToCsv(setId: string): Promise<Blob | null> {
-    let file: Blob | undefined;
-
-    try {
-      file = await lastValueFrom(this.http.get("/api/sets/export/csv/" + setId, { responseType: "blob" }));
-    } catch (e) {
-      return null;
-    }
-
-    return file;
-  }
-
-  /**
-   * Exports set media to a .zip file
-   *
-   * @param setId ID of the set to convert
-   *
-   * @returns Blob of the .zip
-   */
-  async exportSetMedia(setId: string): Promise<Blob | null> {
-    let file: Blob | undefined;
-
-    try {
-      file = await lastValueFrom(this.http.get("/api/sets/export/media/" + setId, { responseType: "blob" }));
-    } catch (e) {
-      return null;
-    }
-
-    return file;
-  }
-
-  /**
-   * Converts a set to an Anki-compatible .apkg file
-   *
-   * @param setId ID of the set to convert
-   *
-   * @returns Blob of the .apkg
-   */
-  async convertSetToApkg(setId: string): Promise<Blob | null> {
-    let file: Blob | undefined;
-
-    try {
-      file = await lastValueFrom(this.http.get("/api/sets/export/anki/" + setId, { responseType: "blob" }));
-    } catch (e) {
-      return null;
-    }
-
-    return file;
-  }
-
-  /**
-   * Converts a set to a .txt which can be imported into Quizlet
-   *
-   * @param setId ID of the set to convert
-   * @param sideDiscriminator Character(s) to discriminate between sides of a card in the string
-   * @param cardDiscriminator Character(s) to discriminate between cards in the string
-   *
-   * @returns Blob of the .txt
-   */
-  async convertSetToQuizletTxt(setId: string, sideDiscriminator: string, cardDiscriminator: string): Promise<Blob | null> {
-    let file: Blob | undefined;
-
-    try {
-      file = await lastValueFrom(
-          this.http.get("/api/sets/export/quizlet/" +
-            setId +
-            "/" +
-            encodeURIComponent(sideDiscriminator) +
-            "/" +
-            encodeURIComponent(cardDiscriminator), { responseType: "blob" })
-      );
-    } catch (e) {
-      return null;
-    }
-
-    return file;
-  }
-
-  /**
    * Creates a set
    *
    * @param body.title Title of the set
@@ -187,76 +102,6 @@ export class SetsService {
         private: body.private,
         cards: body.cards
       }));
-    } catch (e) {
-      return null;
-    }
-
-    if (set.status === ApiResponseOptions.Success) {
-      return set.data;
-    } else return null;
-  }
-
-  /**
-   * Creates a set from an Anki .apkg file
-   *
-   * @param body.title Title of the set
-   * @param body.description Optional, description of the set
-   * @param body.private Whether the set should be publicly visible
-   * @param body.file The .apkg file to be uploaded
-   *
-   * @returns Created `Set` object
-   */
-  async createSetFromApkg(body: {
-    title: string;
-    description?: string;
-    private: boolean;
-    file: File
-  }): Promise<Set | null> {
-    let set: ApiResponse<Set> | undefined;
-
-    const formData = new FormData();
-    formData.append("title", body.title);
-    if (body.description) formData.append("description", body.description);
-    formData.append("private", body.private.toString());
-    formData.append("file", body.file);
-
-    try {
-      set = await lastValueFrom(this.http.post<ApiResponse<Set>>("/api/sets/apkg", formData));
-    } catch (e) {
-      return null;
-    }
-
-    if (set.status === ApiResponseOptions.Success) {
-      return set.data;
-    } else return null;
-  }
-
-  /**
-   * Creates a set from a .csv file
-   *
-   * @param body.title Title of the set
-   * @param body.description Optional, description of the set
-   * @param body.private Whether the set should be publicly visible
-   * @param body.file The .csv file to be uploaded
-   *
-   * @returns Created `Set` object
-   */
-  async createSetFromCsv(body: {
-    title: string;
-    description?: string;
-    private: boolean;
-    file: File
-  }): Promise<Set | null> {
-    let set: ApiResponse<Set> | undefined;
-
-    const formData = new FormData();
-    formData.append("title", body.title);
-    if (body.description) formData.append("description", body.description);
-    formData.append("private", body.private.toString());
-    formData.append("file", body.file);
-
-    try {
-      set = await lastValueFrom(this.http.post<ApiResponse<Set>>("/api/sets/csv", formData));
     } catch (e) {
       return null;
     }
