+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "scholarsome",
-  "version": "1.0.10",
+  "version": "1.1.0",
   "license": "AGPL-3.0-or-later",
   "scripts": {
     "ng": "nx",
